v1/requirements.txt
absl-py==2.1.0
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyascii==0.3.2
anyio==4.2.0
astroid==2.7.3
attmap==0.13.2
attrs==23.2.0
audioread==3.0.1
Babel==2.14.0
bangla==0.0.2
black==24.1.0
blinker==1.7.0
blis==0.7.11
bnnumerizer==0.0.2
bnunicodenormalizer==0.1.6
bokeh==1.4.0
cachetools==5.3.2
catalogue==2.0.10
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
coloredlogs==15.0.1
confection==0.1.4
contourpy==1.2.0
coqpit==0.0.17
coverage==7.4.1
cutlet==0.3.0
cycler==0.12.1
cymem==2.0.8
Cython==3.0.8
dataclasses-json==0.6.4
dateparser==1.1.8
decorator==5.1.1
dimits==0.0.21a0
docopt==0.6.2
eido==0.2.2
einops==0.7.0
encodec==0.1.1
espeak_phonemizer==1.3.1
filelock==3.13.1
Flask==3.0.1
flatbuffers==23.5.26
fonttools==4.47.2
frozenlist==1.4.1
fsspec==2023.12.2
fugashi==1.3.0
g2pkk==0.1.2
google-auth==2.27.0
google-auth-oauthlib==1.2.0
grpcio==1.60.0
gruut==2.2.3
gruut-ipa==0.13.0
gruut-lang-de==2.0.0
gruut-lang-en==2.0.0
gruut-lang-es==2.0.0
gruut-lang-fr==2.0.2
h11==0.14.0
hangul-romanize==0.1.0
httpcore==1.0.2
httpx==0.25.2
huggingface-hub==0.20.3
humanfriendly==10.0
idna==3.6
inflect==7.0.0
isort==5.13.2
itsdangerous==2.1.2
jaconv==0.3.4
jamo==0.4.1
jieba==0.42.1
Jinja2==3.1.3
joblib==1.3.2
jsonlines==1.2.0
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
langchain==0.1.5
langchain-community==0.0.17
langchain-core==0.1.18
langcodes==3.3.0
langsmith==0.0.86
lazy-object-proxy==1.10.0
lazy_loader==0.3
librosa==0.10.1
llvmlite==0.41.1
logmuse==0.2.7
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.4
marshmallow==3.20.2
matplotlib==3.8.2
mccabe==0.6.1
mdurl==0.1.2
mecab-python3==1.0.6
mojimoji==0.0.13
more-itertools==10.2.0
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.4
murmurhash==1.0.10
mutagen==1.47.0
mypy-extensions==1.0.0
networkx==2.8.8
nltk==3.8.1
nose2==0.14.0
num2words==0.5.13
numba==0.58.1
numpy==1.26.3
oauthlib==3.2.2
ollama==0.1.4
onnx==1.15.0
onnxruntime==1.17.0
openai-whisper==20231117
oyaml==1.0
packaging==23.2
pandas==1.5.3
pathspec==0.12.1
peppy==0.40.1
pillow==10.2.0
piper-tts==1.2.0
piper_phonemize @ file:///Users/joshtrim/Development/github/pp
pipestat==0.8.0
platformdirs==4.1.0
pooch==1.8.0
preshed==3.0.9
protobuf==3.20.3
psutil==5.9.8
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyAudio==0.2.14
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
pygame==2.5.2
Pygments==2.17.2
pylint==2.10.2
pynndescent==0.5.11
pyparsing==3.1.1
pypinyin==0.50.0
pysbd==0.3.4
python-crfsuite==0.9.10
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rpds-py==0.17.1
rsa==4.9
safetensors==0.4.2
scikit-learn==1.4.0
scipy==1.12.0
six==1.16.0
smart-open==6.4.0
sniffio==1.3.0
soundfile==0.12.1
soxr==0.3.7
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.25
srsly==2.4.8
SudachiDict-core==20240109
SudachiPy==0.6.8
sympy==1.12
tenacity==8.2.3
tensorboard==2.15.1
tensorboard-data-server==0.7.2
thinc==8.2.2
threadpoolctl==3.2.0
tiktoken==0.5.2
tokenizers==0.15.1
toml==0.10.2
torch==2.1.2
torchaudio==2.1.2
tornado==6.4
tqdm==4.64.0
trainer==0.0.36
transformers==4.37.1
TTS==0.21.3
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.9.0
tzlocal==5.2
ubiquerg==0.7.0
umap-learn==0.5.5
Unidecode==1.3.8
unidic-lite==1.0.8
urllib3==2.1.0
wasabi==1.1.2
weasel==0.3.4
Werkzeug==3.0.1
wrapt==1.12.1
yacman==0.9.3
yarl==1.9.4
